% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseTable.R
\name{baseTable}
\alias{baseTable}
\title{Baseline characteristics table}
\usage{
baseTable(data, by = NULL, time_vars = NULL, include = everything())
}
\arguments{
\item{data}{A data for analysis.}

\item{by}{A categorical variable for group.}

\item{time_vars}{time related variables, which are shown as a median(IQR)}

\item{include}{Variables including in the table. All columns in the data as default.}
}
\value{
A baseline characteristics table with p-value
}
\description{
Create a baseline characteristic table by group variable
}
\examples{
# example code
baseTable(mtcars,
    include = c("mpg", "am","hp","drat","wt","gear","qsec", "disp","cyl"),
    by="cyl")
}
