% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxMultTable.R
\name{coxMultTable}
\alias{coxMultTable}
\title{Mutlivariable Cox regression table}
\usage{
coxMultTable(data, y, time, vars, digits, p.digits)
}
\arguments{
\item{data}{A data for analysis}

\item{y}{Name of a dependent variable or an outcome variable. Should be comprised of 0 and 1.}

\item{time}{A time variable, that is observation duration for the outcome in a survival model.}

\item{vars}{Names of independent variables or response variables.}

\item{digits}{Digits of result values. Default as 2.}

\item{p.digits}{Digits of p-value. Default as 4.}
}
\value{
A table of mutlivariable Cox regression about independent variables included in this analysis.
}
\description{
Perform mutlivariable Cox regressions for numerous variables and create a table
}
