% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxTableCombine.R
\name{coxTableCombine}
\alias{coxTableCombine}
\title{Combine univariable & multivariable Cox regression table}
\usage{
coxTableCombine(data, uni_tbl, mult_tbl)
}
\arguments{
\item{data}{A data for analysis}

\item{uni_tbl}{A data of univariable Cox regression table}

\item{mult_tbl}{A data of multivariable Cox regression table}
}
\value{
A table of merged table of Cox regressions
}
\description{
Perform univariable Cox regressions for numerous variables and create a table
}
