% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxUniTable.R
\name{coxUniTable}
\alias{coxUniTable}
\title{Univariable Cox regression}
\usage{
coxUniTable(data, y, time, Xs)
}
\arguments{
\item{data}{A data for analysis}

\item{y}{Name of a dependent variable or an outcome variable. Should be comprised of 0 and 1.}

\item{time}{A time variable, that is observation duration for the outcome in a survival model.}

\item{Xs}{Names of independent variables or response variables.}
}
\value{
A table of univariable Cox regression about independent variables included in this analysis.
}
\description{
Perform univariable Cox regressions for numerous variables and create a table
}
