% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.cont_mean_sd}
\alias{.cont_mean_sd}
\title{Format for continuous values as mean and sd}
\usage{
.cont_mean_sd(x, digits)
}
\arguments{
\item{x}{The name of variables}
}
\value{
Character values that combined mean and sd.
}
\description{
Internal function to paste mean and sd of a continuous variable.
}
\keyword{internal}
