% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.cont_med_iqr}
\alias{.cont_med_iqr}
\title{Format for continuous values as median and IQR}
\usage{
.cont_med_iqr(x, digits)
}
\arguments{
\item{x}{The name of variables}
}
\value{
Character values that combined median and IQR
}
\description{
Internal function to paste median and iqr of a continuous variable.
}
\keyword{internal}
