% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrMultTable.R
\name{lrMultTable}
\alias{lrMultTable}
\title{Multivariable Logistic regression table}
\usage{
lrMultTable(data, y, Xs)
}
\arguments{
\item{data}{A data for analysis}

\item{y}{Name of a dependent variable or an outcome variable}

\item{Xs}{Names of independent variables or response variables}
}
\value{
A table of multivaraible regression about independent variables included in this analysis
}
\description{
Perform a multivariable logistic regression for numerous variables and create a table
}
\examples{
# example code
lrMultTable(mtcars, 'am', c('mpg','cyl','disp','hp','wt'))
}
