% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrTableCombine.R
\name{lrTableCombine}
\alias{lrTableCombine}
\title{Combine univariable & multivariable Logistic regression table}
\usage{
lrTableCombine(data, uni_tbl, mult_tbl)
}
\arguments{
\item{data}{A data for analysis}

\item{uni_tbl}{A data of univariable logistic regression table}

\item{mult_tbl}{A data of multivariable logistic regression table}
}
\value{
A table of merged table of logistic regressions
}
\description{
Perform univariable logistic regressions for numerous variables and create a table
}
