% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lrUniReg}
\alias{lrUniReg}
\title{Univariable Logistic regression}
\usage{
lrUniReg(data, y, covars)
}
\arguments{
\item{data}{data for analysis}

\item{y}{Name of a Outcome variable}

\item{covars}{Covariates including in analysis}
}
\value{
A Table of Univariable regression
}
\description{
Perform a univariable & multivariable logistic regression
}
