% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrUniTable.R
\name{lrUniTable}
\alias{lrUniTable}
\title{Univariable Logistic regression table}
\usage{
lrUniTable(data, outcome, vars, digits = 2, p.digits = 4)
}
\arguments{
\item{data}{A data for analysis}

\item{outcome}{Name of a dependent variable or an outcome variable.}

\item{vars}{Names of independent variables or response variables.}

\item{digits}{Digits of result values. Default as 2.}

\item{p.digits}{Digits of p-value. Default as 4.}
}
\value{
A table of univariable regression about independent variables included in this analysis
}
\description{
Perform univariable logistic regressions for numerous variables and create a table
}
\examples{
# example code
lrUniTable(mtcars, 'am', c('mpg','cyl','disp','hp','wt'))
}
