% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndist.R
\name{ndist}
\alias{ndist}
\title{Print number of distinct values of variable(s).}
\usage{
ndist(x)
}
\arguments{
\item{x}{Vectors to check number of distinct values.}
}
\description{
Create a table of unique values of selected variables. Compatible with data.table.
}
\examples{
as.data.table(mtcars)[,ndist(cyl)]
as.data.table(mtcars)[,ndist(.SD)]
as.data.table(mtcars)[,lapply(.SD, ndist)]
}
