% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupTable.R
\name{subgroupTable}
\alias{subgroupTable}
\title{Subgroup analysis table}
\usage{
subgroupTable(data, fit, treatment, subgroups, digits = 2)
}
\arguments{
\item{data}{A data for analysis}

\item{fit}{A logistic regression or Cox regression model}

\item{treatment}{A variable that could divide case & control group.}

\item{subgroups}{Categorical Variable(s) for the subgroup analysis.}

\item{digits}{Digits of result values. Default as 2.}
}
\value{
A table of Subgroup analysis with subgroups, level, case, control,  estimates, ci, and p interaction
}
\description{
Perform a subgroup analysis and create a table
}
