% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniTable.R
\name{uniTable}
\alias{uniTable}
\title{Univariable logistic or Cox regression table}
\usage{
uniTable(data, outcome, vars, time = NULL, digits = 2, p.digits = 3)
}
\arguments{
\item{data}{A data for analysis}

\item{outcome}{Name of a dependent variable or an outcome variable}

\item{vars}{Names of independent variables or response variables}

\item{time}{Observation duration variable. If not NULL, cox regression. Default is NULL.}

\item{digits}{Digits of result values. Default as 2.}

\item{p.digits}{Digits of p-value. Default as 4.}
}
\value{
A table of multivariable regression about independent variables included in this analysis
}
\description{
Create a multivariable regression table
}
\examples{
#example code
uniTable(mtcars, 'am', c('mpg','cyl','disp','hp','wt'))

}
