% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varMissTable.R
\name{varMissTable}
\alias{varMissTable}
\title{Missing count, proportion variables table}
\usage{
varMissTable(data, include)
}
\arguments{
\item{data}{A data for the table}

\item{include}{Variables to check the number of missing values}
}
\description{
Create a table that checks the n, proportion of missing values of variables
}
