% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseTable.R
\name{baseTable}
\alias{baseTable}
\title{Baseline characteristics table}
\usage{
baseTable(
  data,
  by = NULL,
  include = NULL,
  median.vars = NULL,
  binary.vars = NULL,
  row.groups = NULL,
  by.order = NULL,
  digits = 1L,
  p.digits = 4L
)
}
\arguments{
\item{data}{A data for analysis.}

\item{by}{A categorical variable for calculating by group. If NULL, all of the data would be aggregated.}

\item{include}{Variables including in the table. If NULL, all variables in the data would be included.}

\item{median.vars}{Variables aggregated to median and IQR.}

\item{binary.vars}{Variables having two categories. These variables will show N and proportions of the last category.}

\item{row.groups}{A group for clinical category. Must be a list with group names of categories and row numbers.}

\item{by.order}{The order of character variable using in `by`. Must be a vector.}

\item{digits}{Digits of result values. Default as 1.}

\item{p.digits}{Digits of p values. Default as 4.}
}
\value{
A baseline characteristics table.
}
\description{
Create a baseline characteristic table by group variable
}
\examples{
# example code
baseTable(mtcars,
 by="cyl",
 include = c("mpg", "am","hp","drat","wt","gear","qsec", "disp"),
 binary.vars = c("vs","am")
 )
}
