% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrMultTable.R
\name{lrMultTable}
\alias{lrMultTable}
\title{Multivariable Logistic regression table}
\usage{
lrMultTable(data, y, vars, digits = 2, p.digits = 4)
}
\arguments{
\item{data}{A data for analysis}

\item{y}{Name of a dependent variable or an outcome variable}

\item{vars}{Names of independent variables or response variables}

\item{digits}{Digits of result values. Default as 2.}

\item{p.digits}{Digits of p-value. Default as 4.}
}
\value{
A table of multivariable regression about independent variables included in this analysis
}
\description{
Perform a multivariable logistic regression for numerous variables and create a table
}
\examples{
# example code
lrMultTable(mtcars, 'am', c('mpg','cyl','disp','hp','wt'))
}
