% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrUniTable.R
\name{lrUniTable}
\alias{lrUniTable}
\title{Univariable Logistic regression table}
\usage{
lrUniTable(data, y, Xs)
}
\arguments{
\item{data}{A data for analysis}

\item{y}{Name of a dependent variable or an outcome variable}

\item{Xs}{Names of independent variables or response variables}
}
\value{
A table of univariable regression about independent variables included in this analysis
}
\description{
Perform univariable logistic regressions for numerous variables and create a table
}
\examples{
# example code
lrUniTable(mtcars, 'am', c('mpg','cyl','disp','hp','wt'))
}
