% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprop.R
\name{nprop}
\alias{nprop}
\title{Print frequency and proportion of a vector}
\usage{
nprop(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector that check frequency and proportion}
}
\description{
Create a table that checks the n, proportion of a vector. Compatible with data.table.
}
